requirements.txt
APScheduler==3.9.1
autoflake==2.0.2
click==8.1.3
colorama==0.4.6
Flask==2.2.3
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
protobuf==3.20.3
pyflakes==3.0.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
tomli==2.0.1
tzdata==2022.7
tzlocal==4.2
Werkzeug==2.2.3
